% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acf_confbands.R
\name{acf_confbands}
\alias{acf_confbands}
\title{Autocorrelation estimation with confidence bands (Time Series, Static Regression)}
\usage{
acf_confbands(
  y,
  H = 25,
  type = "sup-t",
  L = sqrt(length(y)),
  alpha = 0.05,
  plot = TRUE,
  B_hat = NULL,
  B_hat_diag = NULL
)
}
\arguments{
\item{y}{a vector of observations or residuals from a static regression, for which the ACF should be computed.}

\item{H}{maximum number of lags for ACF. The default is \code{25}.}

\item{type}{a character string specifying the type of the confidence band to be computed. Allowed are \code{"sup-t"} (the default), \code{"bonferroni"} or \code{"pointwise"}.}

\item{L}{bandwidth used for covariance estimation. Default is \code{L = sqrt(length(y))}.}

\item{alpha}{significance level. Default is \code{0.05}.}

\item{plot}{logical. If \code{TRUE} (default), a plot of the ACF with confidence bands is provided. If \code{FALSE}, a list of results is provided.}

\item{B_hat}{an optional \verb{H x H} estimated covariance matrix of the autocorrelations. If not provided (the default), the function will estimate the covariance matrix.}

\item{B_hat_diag}{an optional vector of length \code{H} containing the elements of the estimated covariance matrix of the autocorrelations. If not provided (the default), the function will estimate the the necessary elements of the covariance matrix to construct the inference bands. This parameter should only be used in combination with \code{type} set to \code{"bonferroni"} or \code{"pointwise"}. If used with \code{sup-t}, the input provided to \code{B_hat_diag} is ignored and the whole covariance matrix is estimated by the function.}
}
\value{
If \code{plot = TRUE}, a plot of the ACF with confidence bands is provided. If \code{plot = FALSE}, a list with the following elements is returned:
\itemize{
\item{\code{rho_hat} a vector of length \code{H} containing the estimated autocorrelations at each lag.}
\item{\code{conf_band} a matrix of dimension \verb{H x 3} containing the lower bound (\code{lb}), upper bound (\code{ub}) and the width of the confidence band (\code{width}).}
\item{\code{B_hat} or \code{B_hat_diag} estimated \verb{H x H} covariance matrix of autocorrelations for \code{type = "sup_t"} or vector of length \code{H} containing the diagonal elements of covariance matrix for \code{type \%in\% c("bonferroni", "pointwise")} .}
}
}
\description{
Estimate and plot autocorrelations with simultaneous confidence bands of level \code{1-alpha}. Applicable either to a stationary time series or to the residuals from a static regression.
}
\examples{
## Create artificial data from an AR(1) process
 N <- 100 # length of the time series
 I <- 50 # burn-in period
 phi <- 0.5 # coefficient of first lag in AR(1) process
 epsilon <- rnorm(n = N + I, mean = 0, sd = 1) # draw error (I values for burn-in period)
 y <- epsilon[1] # Initialization
 for(tt in 2:(N + I)) y[tt] <- phi*y[tt-1] + epsilon[tt] # AR model
 y <- y[-(1:I)] # disregard burn-in observations
 
 ## Plot autocorrelations with sup-t confidence bands (default)
 acf_confbands(y)
 
 ## Compare to pointwise confidence bands
 acf_confbands(y, type = "pointwise")
 
 ## Obtain results as a list
 res <- acf_confbands(y, plot = FALSE)
 res
 
 

}
