% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acf_sigbands.R
\name{acf_sigbands}
\alias{acf_sigbands}
\title{Autocorrelation estimation with significance bands (Time Series, Static Regression)}
\usage{
acf_sigbands(y, H = 25, type = "simultaneous", alpha = 0.05, plot = TRUE)
}
\arguments{
\item{y}{a vector of observations or residuals from a static regression, for which the ACF should be computed.}

\item{H}{maximum number of lags for ACF. The default is \code{25}.}

\item{type}{a character string specifying the type of the significance band to be computed. Allowed are \code{"simultaneous"} (the default) or \code{"pointwise"}.}

\item{alpha}{significance level. Default is \code{0.05}.}

\item{plot}{logical. If \code{TRUE} (default), a plot of the ACF with confidence bands is provided. If \code{FALSE}, a list of results is provided.}
}
\value{
If \code{plot = TRUE}, a plot of the ACF with significance bands is provided. If \code{plot = FALSE}, a list with the following elements is returned:
\itemize{
\item{\code{rho_hat} a vector of length \code{H} containing the estimated autocorrelations at each lag.}
\item{\code{sig_band} a matrix of dimension \verb{H x 3} containing the lower bound (\code{lb}), upper bound (\code{ub}) and the width of the confidence band (\code{width}).}
}
}
\description{
Estimate and plot autocorrelations with simultaneous significance bands for the null hypothesis of white noise, i.e. the absence of serial correlation. Applicable either to a stationary time series or to the residuals from a static regression. If the sample autocorrelations leave the significance bands at any lag, the null hypothesis is rejected at significance level \code{alpha}.
}
\examples{
## Create artificial data from an AR(1) process
 N <- 100 # length of the time series
 I <- 50 # burn-in period
 phi <- 0.5 # coefficient of first lag in AR(1) process
 epsilon <- rnorm(n = N + I, mean = 0, sd = 1) # draw error (I values for burn-in period)
 y <- epsilon[1] # Initialization
 for(tt in 2:(N + I)) y[tt] <- phi*y[tt-1] + epsilon[tt] # AR model
 y <- y[-(1:I)] # disregard burn-in observations
 
 ## Plot autocorrelations with simultaneous significance bands (default)
 acf_sigbands(y)
 
 ## Compare to pointwise confidence bands
 acf_sigbands(y, type = "pointwise")
 
 ## Obtain results as a list
 res <- acf_sigbands(y, plot = FALSE)
 res
 
 

}
