% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acf_sigbands_dyn.R
\name{acf_sigbands_dyn}
\alias{acf_sigbands_dyn}
\title{Autocorrelation estimation with significance bands for dynamic regressions}
\usage{
acf_sigbands_dyn(
  fit,
  H = 25,
  type = "simultaneous",
  error = "hom",
  alpha = 0.05,
  plot = TRUE,
  Sigma_rho_hat = NULL
)
}
\arguments{
\item{fit}{a fitted model object}

\item{H}{maximum number of lags for ACF. The default is \code{25}.}

\item{type}{a character string specifying the type of the significance band to be computed. Allowed are \code{"simultaneous"} (the default) or \code{"pointwise"}.}

\item{error}{a character string specifying whether conditionally homoskedastic (\code{"hom"}) or heteroskedastic (\code{"het"}) errors should be computed.}

\item{alpha}{significance level. Default is \code{0.05}.}

\item{plot}{logical. If \code{TRUE} (default), a plot of the ACF with confidence bands is provided. If \code{FALSE}, a list of results is provided.}

\item{Sigma_rho_hat}{optional named list containing the estimated covariance matrix under the name \code{estimate}. If not provided (the default), the function will estimate the covariance matrix.}
}
\value{
If \code{plot = TRUE}, a plot of the ACF with significance bands is provided. If \code{plot = FALSE}, a list with the following elements is returned:
\itemize{
\item{\code{rho_hat} a vector of length \code{H} containing the estimated autocorrelations at each lag.}
\item{\code{sig_band} a matrix of dimension \verb{H x 3} containing the lower bound (\code{lb}), upper bound (\code{ub}) and the width of the confidence band (\code{width}).}
\item{\code{Sigma_rho_hat} A list containing
\itemize{
\item \code{estimate}estimated \verb{H x H} covariance matrix of autocorrelations.}
\item{\code{shrinkage_used} logical. If \code{TRUE}, some elements of the plug-in estimator have been set equal to the identity matrix to ensure positive definiteness.}
\item{\code{shrinkage_lag} The lag \code{k}from which on elements of the \verb{H x H} plug-in estimator has been set equal to the identity matrix. If \code{NA}, no shrinkage was necessary to ensure positive definiteness of the plug-in estimator.}
}
}
}
\description{
Estimate and plot autocorrelations with simultaneous significance bands for the null hypothesis of white noise, i.e. the absence of serial correlation. Applicable to the residuals from a dynamic regression with lagged endogenous regressors. If the sample autocorrelations leave the significance bands at any lag, the null hypothesis is rejected at significance level \code{alpha}.
}
\examples{
## Create artificial data from an AR(2) process
 N <- 100 # length of the time series
 I <- 50 # burn-in period
 phi1 <- 0.5 # coefficient of first lag in AR(1) process
 phi2 <- 0.125 # coefficient of second lag in AR(1) process
 epsilon <- rnorm(n = N + I, mean = 0, sd = 1) # draw error
 y <- epsilon[1:2] # Initialization
 for(tt in 3:(N + I)) y[tt] <- phi1*y[tt-1] + phi2*y[tt-2] + epsilon[tt]# AR model
 df <- data.frame(y = y, L1_y = dplyr::lag(y, 1)) # Collect results in a data frame
 df <- df[-(1:I), ] # Disregard burn-in observations
 
 ## Fit an AR(1) model to the data
 fit <- lm(y ~ L1_y, df)
 
 ## Plot autocorrelations of residuals with simultaneous significance bands 
 ## assuming conditionally homoskedastic errors (default)
 acf_sigbands_dyn(fit)
 
 ## Compare to conditionally heteroskedastic errors
 acf_sigbands_dyn(fit, error = "het")
 
 ## Compare to pointwise confidence bands with conditionally homoskedastic errors
 acf_sigbands_dyn(fit, type = "pointwise", error = "hom")
 
 ## Obtain results as a list
 res <- acf_sigbands_dyn(fit, plot = FALSE)
 res
 
 

}
