% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_lags.R
\name{add_lags}
\alias{add_lags}
\title{Helper function: Create lagged variables}
\usage{
add_lags(df, colname, lag_vec)
}
\arguments{
\item{df}{data frame}

\item{colname}{a character string specifiyng the name of the variable we want to lag.}

\item{lag_vec}{vector containing the lags that should be created, i.e. c(1, 4) lags the variable by 1 and by 4.}
}
\value{
the original data frame expandend by columns containing the lagged variable.
}
\description{
Create lagged variables and add it to the data frame.
}
