% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autocor.R
\name{autocor}
\alias{autocor}
\title{Estimate autocorrelations}
\usage{
autocor(y, H)
}
\arguments{
\item{y}{a vector of observations}

\item{H}{maximum number of lags for the ACF}
}
\value{
a vector of length \code{H} containing the estimated autocorrelations at each lag.
}
\description{
Estimate autocorrelations of a given time series for lags \verb{1,...,H}.
}
