% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_Bartlett_true.R
\name{calc_Bartlett_true}
\alias{calc_Bartlett_true}
\title{Calculate the true covariance matrix for a known AR(1) process}
\usage{
calc_Bartlett_true(phi, H)
}
\arguments{
\item{phi}{coefficient of the AR(1) process}

\item{H}{maximum number of lags for the ACF}
}
\value{
a \verb{H x H} covariance matrix
}
\description{
Calculate true covariance matrix according to the Bartlett formula for an AR(1) process using the formulas provided by Cavazos-Cadena (1994).
}
