% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rej.R
\name{calc_rej}
\alias{calc_rej}
\title{Helper function: Calculate frequency of rejections}
\usage{
calc_rej(filename, add_params)
}
\arguments{
\item{filename}{Character string of the filename containing the simulation results.}

\item{add_params}{Parameters to cbind}
}
\value{
A data frame (one row) containing the avaerage over simulations.
}
\description{
Calculate frequency of rejections for significance bands.
}
