% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covar_dynamic.R
\name{covar_dynamic}
\alias{covar_dynamic}
\title{Covariance matrix for dynamic regressions}
\usage{
covar_dynamic(fit, H, error)
}
\arguments{
\item{fit}{a fitted model object}

\item{H}{maximum number of lags for ACF. The default is \code{30}.}

\item{error}{a character string specifying whether conditionally homoskedastic (\code{"hom"}) or heteroskedastic (\code{"het"}) errors should be computed.}
}
\value{
A list containing
\itemize{
\item{\code{estimate} The estimated \verb{H x H} covariance matrix.}
\item{\code{shrinkage_used} logical. If \code{TRUE}, some elements of the plug-in estimator have been set equal to the identity matrix to ensure positive definiteness.}
\item{\code{shrinkage_lag} The lag \code{k}from which on elements of the \verb{H x H} plug-in estimator has been set equal to the identity matrix. If \code{NA}, no shrinkage was necessary to ensure positive definiteness of the plug-in estimator.}
}
}
\description{
Estimate the autocovariance matrix of the residuals from a dynamic regression with lagged endogeneous regressors.
}
