% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equi_quantile.R
\name{equi_quantile}
\alias{equi_quantile}
\title{Equicoordinate quantile}
\usage{
equi_quantile(Sigma, tau)
}
\arguments{
\item{Sigma}{covariance matrix of dimension \verb{H x H}.}

\item{tau}{probability.}
}
\value{
numeric \code{tau} quantile.
}
\description{
Calculate the \code{tau} quantile of \code{M} given a covariance matrix \code{Sigma}.
}
