% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_plot_color.R
\name{make_plot_color}
\alias{make_plot_color}
\title{Helper function: Plot multiple inference bands}
\usage{
make_plot_color(df, segment, line_vec, color_vec)
}
\arguments{
\item{df}{data frame}

\item{segment}{logical. Determines whether sample autocorrelations are plotted are vertically  (\code{TRUE}) or horizontally (\code{FALSE}). The first is used for significance bands, the latter for confidence bands.}

\item{line_vec}{vector containing the linetypes of the different inference bands.}

\item{color_vec}{vector containing the colors of the different inference bands.}
}
\value{
a plot
}
\description{
This helper function is used in the applications to plot several inference bands in one graph using different colors and (possibly) different linetypes.
}
