% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_table.R
\name{make_table}
\alias{make_table}
\title{Helper function: Make a table}
\usage{
make_table(
  df,
  x,
  rnames,
  cnames = c("$H = 1$", "$H = 10 $", "$H = 25$"),
  panel_names = NULL,
  title,
  filename
)
}
\arguments{
\item{df}{data frame}

\item{x}{character specifying the variable of interest}

\item{rnames}{vector containing row names}

\item{cnames}{vector containing column names}

\item{panel_names}{vector containing the panel names}

\item{title}{title of the table}

\item{filename}{write table to the filename}
}
\value{
saves a LaTeX table under the filename
}
\description{
Make table for simulation results.
}
