% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_example.R
\name{plot_example}
\alias{plot_example}
\title{Helper function: Plot analytical examples}
\usage{
plot_example(df, yvar, colorvar, ylabel, hline = NULL)
}
\arguments{
\item{df}{data frame}

\item{yvar}{character string specifying the y-variable.}

\item{colorvar}{character string specifying the group variable.}

\item{ylabel}{label of the y-axis.}

\item{hline}{numeric and optional. If specified, a horizontal line is drawn at that number.}
}
\value{
a plot
}
\description{
This helper function is used to create plots for the analytical examples.
}
