% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_ar1_y.R
\name{sim_ar1_y}
\alias{sim_ar1_y}
\title{DGP: AR(1)}
\usage{
sim_ar1_y(N, I = 50, phi)
}
\arguments{
\item{N}{sample size}

\item{I}{length of burn-in period. Default is 50.}

\item{phi}{slope coefficient}
}
\value{
vector of length \code{N} containing observations generated from an AR(1) process
}
\description{
Generate data according to an AR(1) process.
}
