% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_ar2_fit.R
\name{sim_ar2_fit}
\alias{sim_ar2_fit}
\title{Estimate AR(1) model on data from AR(2) DGP}
\usage{
sim_ar2_fit(N, I = 50, a, phi1, phi2)
}
\arguments{
\item{N}{sample size}

\item{I}{length of burn-in period. Default is 50.}

\item{a}{intercept}

\item{phi1}{coefficient of lag 1}

\item{phi2}{coefficient of lag 2}
}
\value{
a fitted model object
}
\description{
Generate data according to an AR(2) process and fit an AR(1) model to it.
}
