% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_plot.R
\name{make_plot}
\alias{make_plot}
\title{Helper function: Plot autocorrelations with inference bands}
\usage{
make_plot(color, segment, H, rho_hat, lb, ub)
}
\arguments{
\item{color}{color of the lines.}

\item{segment}{logical. Determines whether sample autocorrelations are plotted are vertically  (\code{TRUE}) or horizontally (\code{FALSE}). The first is used for significance bands, the latter for confidence bands.}

\item{H}{maximum number of lags for ACF.}

\item{rho_hat}{A vector of length \code{H} containing the sample autocorrelations.}

\item{lb}{A vector of length \code{H} containing the lower bound of the inference band.}

\item{ub}{A vector of length \code{H} containing the upper bound of the inference band.}
}
\value{
a vector of length \code{H} containing the estimated autocorrelations at each lag.
}
\description{
This helper function is used in the three main functions to create a plot.
}
